refactor(agent): extract R2DLearner; launcher owns the encoder (decision 1)
The agent internals move verbatim to src/r2dreamer/learner.py as
R2DLearner, which receives the trainable encoder module INJECTED
(encoder + encoder_init_obs constructor args) and never constructs one.
The RNG key-split discipline in __init__ is unchanged, so a given seed
keeps producing bit-identical parameters.

launch/train.py becomes the composition root per IDEA decisions 1 + 5:
recipe.build_module(cfg) builds the encoder, the first prepared frame
from collector.reset() defines the obs spec AND serves as the enc.init
dummy, check_branch_keys runs against the real frame for composite
encoders, and the launcher injects the encoder into the learner. The
launcher no longer imports R2DreamerAgent.

agent.py shrinks to the temporary compatibility shim (DELETIONS.md):
R2DreamerAgent(cfg, rng) = recipe lookup + init-dummy + R2DLearner, plus
the from_checkpoint composition path; deleted in the final step.

Gates: full CPU suite 586 passed / 0 failed; pylint 345 msgs 9.54/10
(= baseline); golden runs byte-identical with properly isolated module
resolution (clean base worktree @8065e20 vs branch, PYTHONPATH pinned:
cnn de54e634..., hybrid 3a28cda0... on both sides).
